refactor: contain chat stream session
